refactor: move count_by to separate file
